==> notebook/01_data_preprocessing_with_spell.ipynb <==
import re
import pandas as pd
from langdetect import detect
import nltk
import spacy
import string
import os

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize


import os

current_path = os.getcwd()
print("Current path:", current_path)


custom_directory = "../artifacts"

nltk.download('punkt', download_dir=custom_directory)
nltk.download('stopwords', download_dir=custom_directory)


#change the nltk base directory
nltk.data.path.append(custom_directory)

# Going to read the csv file

df = pd.read_csv("../artifacts/sample_sentiment_analysis.csv")

df

df = df[["Comment","Rating"]]

df

# filtering the english review

!pip install joblib

<h3>import joblib for parallel processing</h3>
<p>use this for use all the cpu cores and threds</p>

from joblib import Parallel, delayed

# definition for english language detect

def is_english(text):
    try:
        return detect(text) == 'en'
    except:
        return False

# use the joblib parallel "is_english" function

english_flags = Parallel(n_jobs=-1)(
    delayed(is_english)(text) for text in df['Comment']
)

# remove the other language and filter the english

df['is_english'] = english_flags
df = df[df['is_english']].drop(columns='is_english').reset_index(drop=True)

df

# import the symspellpy

from symspellpy import SymSpell,Verbosity

# create spell check function

sym_spell = SymSpell(max_dictionary_edit_distance=2, prefix_length=7)
sym_spell.load_dictionary("frequency_dictionary_en_82_765.txt", term_index=0, count_index=1)

def correct_spelling(text):
    suggestions = sym_spell.lookup_compound(text, max_edit_distance=2)
    return suggestions[0].term if suggestions else text

# df['Corrected'] = df['Comment'].apply(correct_spelling)

spell_checked = Parallel(n_jobs=-1)(
    delayed(correct_spelling)(text) for text in df['Comment'])

df['Corrected'] = spell_checked

# make solid copy to protect the work

df_copy =df.copy()

 # remove the stop words using joblib for fast

stop_words = set(stopwords.words('english'))
def clean_text(text):
    # lowercase
    text = text.lower()
    # remove URLs
    text = re.sub(r"http\S+|www\S+|https\S+", '', text, flags=re.MULTILINE)
    # remove punctuation
    text = text.translate(str.maketrans('', '', string.punctuation))
    # remove numbers
    text = re.sub(r'\d+', '', text)
    # remove extra spaces
    text = re.sub(r'\s+', ' ', text).strip()
    return text

df['cleaned_text']  = Parallel(n_jobs=-1)(delayed(clean_text)(text) for text in df['Corrected'])

df

# using lemmatize text
<p>for more accuracy we use lemmatize text instead of stemming</p>
<p>nlp.pipe(n_process=…)is	Designed for spaCy for safe parallelism and it  keeps order	Only works with spaCy</p>

nlp = spacy.load("en_core_web_sm")

def lemmatize_doc(doc):
    return ' '.join([
        token.lemma_
        for token in doc
        if token.is_alpha and not token.is_stop
    ])

# Use nlp.pipe to process docs in parallel
docs = nlp.pipe(df['cleaned_text'], n_process=-1)

# Apply lemmatization
df['lemmatized_text'] = [lemmatize_doc(doc) for doc in docs]


df.to_csv("../artifacts/sample_checked.csv", index=True)

df
