==> notebook/02.1_model bulding.ipynb <==
import pandas as pd
import numpy as np
import os

df= pd.read_csv('../artifacts/sample_checked.csv')

df.head()

df = df[["lemmatized_text","Rating"]]

df.head()

from collections import Counter

vocab = Counter()

# Option 1: Replace non-strings with empty strings
df['lemmatized_text'] = df['lemmatized_text'].astype(str)

# Option 2: Drop rows where 'lemmatized_text' is NaN
df = df.dropna(subset=['lemmatized_text'])


### create vocabulary

for sentence in df['lemmatized_text']:
    vocab.update(sentence.split())

len(vocab)

df.shape

### trim the vocabulary

tokens= [key for key in vocab if vocab[key]>10]

len(tokens)

def save_vocabulary(lines,filename):
    data = '\n'.join(lines)
    with open(filename, 'w', encoding="utf-8") as f:
        f.write(data)

save_vocabulary(tokens,'../artifacts/vocabulary.txt')

### split the dataset to train and validate

X = df["lemmatized_text"]
Y = df["Rating"]

from sklearn.model_selection import train_test_split

X_train, X_test, Y_train, Y_test = train_test_split(X, Y, test_size=0.2, shuffle=True)


X_test.shape

X_train.shape

### vectorization

def vectorization(dataset,vocabulary):
    vectorized_dataset = []
    for sentence in dataset:
        sentence_lst = np.zeros(len(vocabulary));
        for i in range(len(vocabulary)):
            if vocabulary[i] in sentence.split():
                sentence_lst[i]= 1

        vectorized_dataset.append(sentence_lst)
    vectorized_dataset_new = np.asarray(vectorized_dataset,dtype=np.float32)
    return vectorized_dataset_new

vectorized_x_train = vectorization(X_train,tokens)

vectorized_x_test = vectorization(X_test,tokens)

Y_train.value_counts()
